Traffic control agorithm
This algorithm takes scaled traffic density values for each lane which are determined by image processing at traffic signals and assigns signal STOP/GO time and signal priority based on the density values.
Higher the priority, the more the traffic in the lane. Hence, more time should be alloted to that lane.